% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropls_helpers.R
\name{get_loadings}
\alias{get_loadings}
\title{Get axis loadings from models created by `ropls::opls()`}
\usage{
get_loadings(.model)
}
\arguments{
\item{.model}{a pls object created by \code{\link{opls}}}
}
\value{
a tibble
}
\description{
Provides a wrapper for \code{\link{getLoadingMN}} from the \link{ropls} package that returns a tibble rather than a matrix
}
\examples{
\dontrun{
pls.model <- opls(X, Y)
get_loadings(pls.model)
}
}
