% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropls_helpers.R
\name{get_modelinfo}
\alias{get_modelinfo}
\title{Retrieve model parameters from models created by `ropls::opls()`
For PCA, returns percent variance explained by each axis.  For (o)PLS(-DA), returns variance explained by axes and cross-validation statistics.}
\usage{
get_modelinfo(model)
}
\arguments{
\item{model}{a model object created by `opls()`}
}
\value{
a list of two dataframes, `axis_stats` and `validation`
}
\description{
Retrieve model parameters from models created by `ropls::opls()`
For PCA, returns percent variance explained by each axis.  For (o)PLS(-DA), returns variance explained by axes and cross-validation statistics.
}
\examples{
\dontrun{
pls.model <- opls(X, Y)
get_modelinfo(pls.model)
}
}
