% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropls_helpers.R
\name{get_plotdata}
\alias{get_plotdata}
\title{Extract data for plotting (O)PLS(-DA) data with ggplot2}
\usage{
get_plotdata(model)
}
\arguments{
\item{model}{An object created by \code{\link{opls}}}
}
\value{
A list containing dataframes for scores, loadings, axis statistics (%variance explained), and model cross-validation
}
\description{
Extracts relevant data from an "opls" object for making annotated score plots with ggplot2 or other plotting packages.
}
\examples{
\dontrun{
library(ropls)
data(sacurine)
sacurine.oplsda <- opls(sacurine$dataMatrix, sacurine$sampleMetadata[, "gender"],
                        predI = 1,
                        orthoI = NA)
df <- get_plotdata(sacurine.oplsda)
}
}
