% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropls_helpers.R
\name{get_scores}
\alias{get_scores}
\title{Get axis scores from models created by `ropls::opls()`
Returns a dataframe of PC axis scores for PCA, predictive axis scores for PLS and PLS-DA, and predictive and orthogonal axis scores for OPLS and OPLS-DA models.}
\usage{
get_scores(model)
}
\arguments{
\item{model}{a model object created by `opls()`}
}
\value{
a dataframe
}
\description{
Get axis scores from models created by `ropls::opls()`
Returns a dataframe of PC axis scores for PCA, predictive axis scores for PLS and PLS-DA, and predictive and orthogonal axis scores for OPLS and OPLS-DA models.
}
\examples{
\dontrun{
pls.model <- opls(X, Y)
get_scores(pls.model)
}
}
