% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropls_helpers.R
\name{get_VIP}
\alias{get_VIP}
\title{Get VIP scores from PLS and OPLS models created by `ropls::opls()`}
\usage{
get_VIP(.model)
}
\arguments{
\item{.model}{a pls object created by \code{\link{opls}}}
}
\value{
a tibble
}
\description{
Provides a wrapper for \code{\link{getVipVn}} from the \link{ropls} package that returns a tibble rather than a named numeric vector.
}
\examples{
\dontrun{
pls.model <- opls(X, Y)
get_VIP(pls.model)
}
}
