% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holodeck-package.R
\docType{package}
\name{holodeck-package}
\alias{holodeck}
\alias{holodeck-package}
\title{holodeck: A Tidy Interface for Simulating Multivariate Data}
\description{
Provides pipe-friendly (%>%) wrapper functions for MASS::mvrnorm() to create simulated multivariate data sets with groups of variables with different degrees of variance, covariance, and effect size.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Aariq/holodeck}
  \item Report bugs at \url{https://github.com/Aariq/holodeck/issues}
}

}
\author{
\strong{Maintainer}: Eric Scott \email{scottericr@gmail.com} (\href{https://orcid.org/0000-0002-7430-7879}{ORCID})

}
\keyword{internal}
