% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{holodeck}
\alias{holodeck}
\alias{holodeck-package}
\title{holodeck: A package for simulating multivariate datasets}
\description{
The `holodeck` package contains functions for creating "chunks" of variables with different degrees of co-variance (collinearity) and discrimination among groups (i.e. levels of a categorical variable).
}
\details{
What make it 'tidy'?  All `sim_*` functions accept dataframes or tibbles as their first argument and return tibbles, meaning they work with the pipe operator (`%>%`) from the `dplyr` package.
}
