% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{:=}
\alias{:=}
\title{Definition operator}
\arguments{
\item{x}{An object to test.}

\item{lhs, rhs}{Expressions for the LHS and RHS of the definition.}
}
\description{
Internally, this package uses the definition operator, \code{:=},
to make assignments that require computing on the LHS.
}
