% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropls_plots.R
\name{plot_oplsda}
\alias{plot_oplsda}
\title{Plot OPLS-DA models produced by `ropls::opls()`}
\usage{
plot_oplsda(ropls_pls, annotate = c("caption", "subtitle"))
}
\arguments{
\item{ropls_pls}{an OPLS-DA model with a discrete Y variable produced by `ropls::opls()`}

\item{annotate}{location on the plot to print model statistics}
}
\value{
a ggplot object
}
\description{
Plot OPLS-DA models produced by `ropls::opls()`
}
\examples{
\dontrun{
plot_oplsda(oplsda)
}
}
