% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropls_plots.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA models created by `ropls::opls()`}
\usage{
plot_pca(ropls_pca, group_var = NULL, annotate = c("caption",
  "subtitle", "none"))
}
\arguments{
\item{ropls_pca}{a PCA model produced by `ropls::opls()`}

\item{group_var}{a discrete variable used to plot groups}

\item{annotate}{location on the plot to print model statistics}
}
\value{
a ggplot object
}
\description{
Plot PCA models created by `ropls::opls()`
}
\examples{
\dontrun{
plot_pca(pca, data$treatment)
}
}
