% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_diag}
\alias{set_diag}
\title{Pipe friendly wrapper to `diag(x) <- value`}
\usage{
set_diag(x, value)
}
\arguments{
\item{x}{a matrix}

\item{value}{either a single value or a vector of length equal to the diagonal of `x`.}
}
\value{
a matrix
}
\description{
Pipe friendly wrapper to `diag(x) <- value`
}
\examples{
library(dplyr)
matrix(0,3,3) \%>\%
set_diag(1)
}
