% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_multvar.R
\name{sim_covar}
\alias{sim_covar}
\title{Simulate co-varying variables}
\usage{
sim_covar(.data = NULL, N = NULL, p, var, cov, name = NA,
  seed = NA)
}
\arguments{
\item{.data}{An optional dataframe.  If a dataframe is supplied, simulated categorical data will be added to the dataframe. Either `.data` or `N` must be supplied.}

\item{N}{Total number of observations/rows to simulate if `.data` is not supplied.}

\item{p}{number of variables to simulate.}

\item{var}{variance used to construct variance-covarinace matrix.}

\item{cov}{covariance used to construct variance-covarinace matrix.}

\item{name}{an optional name to be appended to the column names in the output.}

\item{seed}{an optional seed for random number generation.  If `NA` (default) a random seed will be used.}
}
\value{
a tibble
}
\description{
Adds a group of variables (columns) with a given variance and covariance to a data frame or tibble
}
\examples{
library(dplyr)
sim_cat(N = 30, n_groups = 3) \%>\%
sim_covar(p = 5, var = 1, cov = 0.5, name = "correlated")
}
