% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_multvar2.R
\name{sim_df}
\alias{sim_df}
\title{Simulate a dataset with a given number of observations and groups}
\usage{
sim_df(N, n_groups)
}
\arguments{
\item{N}{Total number of observations/rows}

\item{n_groups}{How many groups or treatments to simulate}
}
\value{
a tibble
}
\description{
This is a simple wrapper that creates a tibble of length `N` with a single column `groups`.  It will warn if there are fewer than three replicates per group.
}
\examples{
df <- sim_df(30, 3)
}
