% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_multvar2.R
\name{sim_missing}
\alias{sim_missing}
\title{Simulate missing values}
\usage{
sim_missing(df, prop, seed = NA)
}
\arguments{
\item{df}{a dataframe}

\item{prop}{proportion of values to be set to NA}
}
\value{
a dataframe with NAs
}
\description{
Simulate missing values
}
\examples{
library(dplyr)
df <- sim_df(10, 2) \%>\% 
sim_covar(10, 1, 0.5) \%>\% 
sim_missing(0.05)
}
