% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropls_plots.R
\name{plot_opls}
\alias{plot_opls}
\title{Plot OPLS regression models produced by `ropls::opls()`}
\usage{
plot_opls(ropls_pls, annotate = c("caption", "subtitle"))
}
\arguments{
\item{ropls_pls}{a PLS model with a discrete Y variable produced by `ropls::opls()`}

\item{annotate}{place to put model statistics on the plot}
}
\value{
a ggplot object
}
\description{
Plot OPLS regression models produced by `ropls::opls()`
}
\examples{
\dontrun{
plot_opls(opls)
}
}
