% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropls_plots.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA models created by `ropls::opls()`}
\usage{
plot_pca(ropls_pca, group_var = NULL, annotate = c("caption",
  "subtitle"))
}
\arguments{
\item{annotate}{place to put model statistics on the plot}
}
\value{
a ggplot object
}
\description{
Plot PCA models created by `ropls::opls()`
}
\examples{
\dontrun{
plot_pca(pca, data$treatment)
}
}
