% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_multvar.R
\name{sim_cat}
\alias{sim_cat}
\title{Simulate a dataset with a given number of observations and groups}
\usage{
sim_cat(N, n_groups, name = "group")
}
\arguments{
\item{N}{Total number of observations/rows}

\item{n_groups}{How many groups or treatments to simulate}

\item{name}{The column name for the grouping variable.  Defaults to "group".}
}
\value{
a tibble
}
\description{
This is a simple wrapper that creates a tibble of length `N` with a single column `groups`.  It will warn if there are fewer than three replicates per group.
}
\details{
To-do: make this optionally accpet a dataframe and add categorical variables
Make this optionally create multiple categorical variables as being nested or crossed or random
}
\examples{
df <- sim_cat(30, 3)
}
