% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_multvar.R
\name{sim_covar}
\alias{sim_covar}
\title{Simulate co-varying variables}
\usage{
sim_covar(df, p, var, cov, name = NA, seed = NA)
}
\arguments{
\item{df}{a data frame or tibble}

\item{p}{number of variables to simulate}

\item{var}{variance used to construct variance-covarinace matrix.}

\item{cov}{covariance used to construct variance-covarinace matrix.}

\item{name}{an optional name to be appended to the column names in the output}

\item{seed}{an optional seed for random number generation.  If `NA` (default) a random seed will be used.}
}
\value{
a tibble
}
\description{
Simulate co-varying variables
}
\examples{
df <- sim_cat(30, 3)
sim_covar(df, p = 5, var = 1, cov = 0.5, name = "correlated")
}
