% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_multvar.R
\name{sim_missing}
\alias{sim_missing}
\title{Simulate missing values}
\usage{
sim_missing(df, prop, seed = NA)
}
\arguments{
\item{df}{a dataframe}

\item{prop}{proportion of values to be set to NA}

\item{seed}{an optional seed for random number generation.  If `NA` a random seed will be used.}
}
\value{
a dataframe with NAs
}
\description{
Takes a data frame and randomly replaces a user-supplied proportion of values with `NA`.
}
\examples{
library(dplyr)
df <- sim_cat(10, 2) \%>\%
sim_covar(10, 1, 0.5) \%>\%
sim_missing(0.05)
}
